+++ packages/bridge-controller/src/utils/metrics/constants.ts
@@ -50,6 +50,11 @@ export enum MetaMetricsSwapsEventSource {
   TokenView = 'Token View',
   TrendingExplore = 'Trending Explore',
   Rewards = 'Rewards',
+  ActivityTabEmptyState = 'Activity Tab Empty State',
+  TransactionShield = 'Transaction Shield',
+  TransactionDetails = 'Transaction Details',
+  DeepLink = 'Deep Link',
+  Unknown = 'Unknown',
 }
 
 export enum InputAmountPreset {